feat: restructure project into backend directory and DDD layout (resolves #2)
