+++ backend/Makefile
@@ -13,10 +13,10 @@ build:
 	go build -o bin/api cmd/api/main.go
 
 migrate-up:
-	migrate -path migrations -database "$(DB_URL)" up
+	migrate -path db/migrations -database "$(DB_URL)" up
 
 migrate-down:
-	migrate -path migrations -database "$(DB_URL)" down
+	migrate -path db/migrations -database "$(DB_URL)" down
 
 test:
 	go test ./...
