+++ backend/pkg/utils/google_auth.go
@@ -1,4 +1,4 @@
-package security
+package utils
 
 import (
 	"context"
