+++ backend/pkg/utils/jwt.go
@@ -1,4 +1,4 @@
-package security
+package utils
 
 import (
 	"fmt"
