+++ backend/pkg/utils/password.go
@@ -1,4 +1,4 @@
-package security
+package utils
 
 import (
 	"golang.org/x/crypto/bcrypt"
